% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cossim.R
\name{cossim}
\alias{cossim}
\title{Compute cosine similarity among rows in a matrix}
\usage{
cossim(x)
}
\arguments{
\item{x}{A matrix of examples (rows) by features (columns)}
}
\value{
A matrix with dimensions `nrow(x)` by `nrow(x)`.
}
\description{
Compute cosine similarity among rows in a matrix
}
