% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cossim.R
\name{simple_cossim}
\alias{simple_cossim}
\title{Simple, but inefficient, implementation of cosine similarity for testing}
\usage{
simple_cossim(x)
}
\arguments{
\item{x}{A matrix of examples (rows) by features (columns)}
}
\value{
A matrix with dimensions `nrow(x)` by `nrow(x)`.
}
\description{
Simple, but inefficient, implementation of cosine similarity for testing
}
